# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Receipt Tracker API
  version: 0.1.0
  description: >
    API contract for a mobile-first receipt tracking app with OCR normalization,
    review, reporting, and outbound sync to QuickBooks Online and Microsoft Excel.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    User:
      type: object
      required: [id, email, defaultCurrency, timezone]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
          nullable: true
        defaultCurrency:
          type: string
          minLength: 3
          maxLength: 3
        timezone:
          type: string
    SessionExchangeRequest:
      type: object
      required: [provider, idToken]
      properties:
        provider:
          type: string
          enum: [google, apple, microsoft]
        idToken:
          type: string
        deviceName:
          type: string
          nullable: true
    SessionExchangeResponse:
      type: object
      required: [accessToken, refreshToken, expiresInSeconds, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSeconds:
          type: integer
        user:
          $ref: '#/components/schemas/User'
...
    Category:
      type: object
      required: [id, name, source, isActive]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        source:
          type: string
          enum: [system, user, quickbooks]
        externalAccountId:
          type: string
          nullable: true
        externalAccountType:
          type: string
          nullable: true
        isActive:
          type: boolean
    Vendor:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        normalizedName:
          type: string
        lastUsedCategoryId:
          type: string
          format: uuid
          nullable: true
    ConnectUrlRequest:
      type: object
      required: [redirectUri]
      properties:
        redirectUri:
          type: string
          format: uri
    ConnectUrlResponse:
      type: object
      required: [authorizationUrl, state]
      properties:
        authorizationUrl:
          type: string
          format: uri
        state:
          type: string
    OAuthCallbackRequest:
      type: object
      required: [code, state]
      properties:
        code:
          type: string
        state:
          type: string
        realmId:
          type: string
          nullable: true
    IntegrationStatus:
      type: object
      required: [provider, status]
      properties:
        provider:
          type: string
          enum: [quickbooks, microsoft]
        status:
          type: string
          enum: [not_connected, connected, expired, revoked, error]
        connectedAt:
          type: string
          format: date-time
          nullable: true
        externalTenantName:
          type: string
          nullable: true
        scopes:
          type: array
          items:
            type: string
...
paths:
...
  /auth/session/exchange:
    post:
      tags: [auth]
      summary: Exchange an external identity token for an app session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionExchangeRequest'
      responses:
        '200':
          description: Session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionExchangeResponse'
        '401':
          description: Invalid external token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [auth]
      summary: Get the current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /categories:
    get:
      tags: [categories]
      summary: List categories available to the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
  /vendors:
    get:
      tags: [vendors]
      summary: List known vendors for the current user
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Vendor list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vendor'
  /integrations/quickbooks/connect-url:
    post:
      tags: [integrations]
      summary: Create a QuickBooks authorization URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectUrlRequest'
      responses:
        '200':
          description: Authorization URL created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectUrlResponse'
  /integrations/quickbooks/callback:
    post:
      tags: [integrations]
      summary: Exchange the QuickBooks auth code for encrypted tokens
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthCallbackRequest'
      responses:
        '200':
          description: Integration connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationStatus'
  /integrations/quickbooks/status:
    get:
      tags: [integrations]
      summary: Get the current QuickBooks connection status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: QuickBooks status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationStatus'
  /integrations/microsoft/connect-url:
    post:
      tags: [integrations]
      summary: Create a Microsoft authorization URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectUrlRequest'
      responses:
        '200':
          description: Authorization URL created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectUrlResponse'
  /integrations/microsoft/callback:
    post:
      tags: [integrations]
      summary: Exchange the Microsoft auth code for encrypted Graph tokens
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthCallbackRequest'
      responses:
        '200':
          description: Integration connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationStatus'
  /integrations/microsoft/status:
    get:
      tags: [integrations]
      summary: Get the current Microsoft connection status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Microsoft status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationStatus'
...